rename `mb_ro_on_layout` → `reorder`
